Added multiple display types
Added the display_type config variable that allows switching between
three display types: index-url-tags (default), index-title-tags,
index-title-tags plus the url on the next line.

Regardless of display type, the url and tags are always editable. Added
the ability to edit the title of the bookmark.

Added the max_str_width that controls how long titles and urls can be
(default 80).